+++ torch/csrc/jit/passes/frozen_conv_add_relu_fusion.h
@@ -6,6 +6,9 @@
 namespace torch {
 namespace jit {
 
+TORCH_API extern std::function<void(std::shared_ptr<Graph>&)>
+    _fuseFrozenConvAddReluImpl;
+
 TORCH_API void FuseFrozenConvAddRelu(std::shared_ptr<Graph>& graph);
 
 } // namespace jit
